Werkt!
Hij werkt :D Iedere keer na code generation moet je wel de DR aanpassen in Middleware->LoRaWAN->third party->Region->EU868.h Tx Default en minimum op 3.

+++ Core/Inc/applicatie.h
@@ -21,7 +21,8 @@
 //======================================================================================================
 float Sensor_read();
 void Sensor_Init(void);
-uint16_t Sensor_Data(void);
+//uint16_t Sensor_Data(void);
+float Sensor_Data(void);
 void one_shot_trigger(void);
 
 void I2C_scan(void);
